Create "index.html" And Edit "README.md" initially
Create the "index.html" file where all the other
which acts as a homepage containing links to
other pages of the site. Fill the boilerplate
code to start the project. Add details to the
"README.md" file describing the overview of the
project details. Edit "README.md" file later
whenever needed.