Create "recipes" Folder And "pancake.html" In It
Create a "recipes" folder to contain various recipes. Create the first
recipe file, "pancake.html" to begin with. Fill in the boilerplate code
in "pancake.html" and add the link to the "index.html" in the body.
Add the link to "pancake.html" in the "index.html" for navigating to it.