Add recipe making instructions to "pancake.html"
Create an "images" folder to store the sample photo of pancakes.
Add a link to the pancake photo to "pancake.html".  Add a description
about the  dish below the photo along with it's name as the heading.
Deploy the usage of unordered lists andordered lists to represent
ingredients and dish-making instructions respectively. Add the creator
of the photo's profile to "README.md".